Various refactorings reg. review comments
