+++ src/ibex_bluesky_core/callbacks/__init__.py
@@ -1,0 +1,1 @@
+"""For callbacks that relate to the BlueSky run engine."""
